Moved code to midi.rs
